[App configuration] rename application
