+++ analysis_options.yaml
@@ -1,68 +1,45 @@
-# This file configures the analyzer, which statically analyzes Dart code to
-# check for errors, warnings, and lints.
-#
-# The issues identified by the analyzer are surfaced in the UI of Dart-enabled
-# IDEs (https://dart.dev/tools#ides-and-editors). The analyzer can also be
-# invoked from the command line by running `flutter analyze`.
-
-# The following line activates a set of recommended lints for Flutter apps,
-# packages, and plugins designed to encourage good coding practices.
 include: package:flutter_lints/flutter.yaml
 
 analyzer:
-  errors:
-    invalid_annotation_target: ignore
-  plugins:
-    - dart_code_metrics
-  exclude: 
-    - "**/*.mocks.dart"
-    - "**/**.g.dart"
-    - "**/**.freezed.dart"
-    - lib/l10n/**
+   errors:
+     invalid_annotation_target: ignore
+   plugins:
+     - dart_code_metrics
+   exclude:
+     - "**/*.mocks.dart"
+     - "**/**.g.dart"
+     - "**/**.freezed.dart"
+     - lib/l10n/**
 
 linter:
-  # The lint rules applied to this project can be customized in the
-  # section below to disable rules from the `package:flutter_lints/flutter.yaml`
-  # included above or to enable additional rules. A list of all available lints
-  # and their documentation is published at
-  # https://dart-lang.github.io/linter/lints/index.html.
-  #
-  # Instead of disabling a lint rule for the entire project in the
-  # section below, it can also be suppressed for a single line of code
-  # or a specific dart file by using the `// ignore: name_of_lint` and
-  # `// ignore_for_file: name_of_lint` syntax on the line or in the file
-  # producing the lint.
   rules:
-    # avoid_print: false  # Uncomment to disable the `avoid_print` rule
-    # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
-
-# Additional information about this file can be found at
-# https://dart.dev/guides/language/analysis-options
+    avoid_print: false
+    prefer_single_quotes: true
 
 dart_code_metrics:
-  anti-patterns:
-    - long-method
-    - long-parameter-list
-  metrics:
-    cyclomatic-complexity: 20
-    maximum-nesting-level: 5
-    number-of-parameters: 4
-    source-lines-of-code: 50
-  metrics-exclude:
-    - test/**
-  rules:
-    - newline-before-return
-    - no-boolean-literal-compare
-    - no-empty-block
-    - prefer-trailing-comma
-    - prefer-conditional-expressions
-    - no-equal-then-else
-  rules-exclude:
-    - test/**
-    - lib/**.freezed.dart
-    - lib/**.g.dart
-    - no-equal-arguments:
-        exclude:
-          - lib/**.freezed.dart
-          - lib/**.g.dart
-          - test/**
+   anti-patterns:
+     - long-method
+     - long-parameter-list
+   metrics:
+     cyclomatic-complexity: 20
+     maximum-nesting-level: 5
+     number-of-parameters: 4
+     source-lines-of-code: 50
+   metrics-exclude:
+     - test/**
+   rules:
+     - newline-before-return
+     - no-boolean-literal-compare
+     - no-empty-block
+     - prefer-trailing-comma
+     - prefer-conditional-expressions
+     - no-equal-then-else
+   rules-exclude:
+     - test/**
+     - lib/**.freezed.dart
+     - lib/**.g.dart
+     - no-equal-arguments:
+         exclude:
+           - lib/**.freezed.dart
+           - lib/**.g.dart
+           - test/**
