+++ android/app/src/main/kotlin/com/aborovskoy/flutter_shop/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.store_app
+package com.aborovskoy.flutter_shop
 
 import io.flutter.embedding.android.FlutterActivity
