+++ pubspec.yaml
@@ -1,8 +1,7 @@
-name: store_app
+name: flutter_shop
 description: A new Flutter project.
 version: 1.0.0+1
-publish_to: 'none' 
-
+publish_to: 'none'
 
 environment:
   sdk: ">=2.16.2 <3.0.0"
@@ -28,12 +27,11 @@ dependencies:
 dev_dependencies:
   build_runner: ^2.1.11
   dart_code_metrics: ^4.15.2
-  flutter_lints: ^1.0.0
+  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
   freezed: ^2.0.3+1
   mockito: ^5.2.0
-
 
 flutter:
   uses-material-design: true
@@ -41,7 +39,7 @@ flutter:
 
   assets:
     - assets/images/product_placeholder.png
-  
+
   fonts:
     - family: Lato
       fonts:
