+++ test/common/data/repository/orders_repository_test.dart
@@ -2,8 +2,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 
-import 'package:store_app/common/data/model/order_item.dart';
-import 'package:store_app/common/orders/repository/orders_repository.dart';
+import 'package:flutter_shop/common/data/model/order_item.dart';
+import 'package:flutter_shop/common/orders/repository/orders_repository.dart';
 
 import 'orders_repository_test.mocks.dart';
