+++ test/common/data/repository/orders_repository_test.mocks.dart
@@ -1,14 +1,14 @@
 // Mocks generated by Mockito 5.2.0 from annotations
-// in store_app/test/common/data/repository/orders_repository_test.dart.
+// in flutter_shop/test/common/data/repository/orders_repository_test.dart.
 // Do not manually edit this file.
 
 import 'dart:async' as _i3;
 
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:store_app/common/data/model/cart_item.dart' as _i5;
-import 'package:store_app/common/data/model/order_item.dart' as _i4;
-import 'package:store_app/common/orders/repository/orders_repository.dart'
+import 'package:flutter_shop/common/data/model/cart_item.dart' as _i5;
+import 'package:flutter_shop/common/data/model/order_item.dart' as _i4;
+import 'package:flutter_shop/common/orders/repository/orders_repository.dart'
     as _i2;
+import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
